// Front/src/app/actividade/model/funcionario.ts
export class Funcionario {

  //users
  email: string;
  username:string;

  //Pessoa
  id: number;
  nome: string;
  ndi: string;
  user_id: number;
  genero: string;
  estado_civil: string;
  tipo_identificacao: string;
  data_nascimento: any;
  base_municipio_id: number;
  endereco: string;
  data_emissao: any;
  banco: string;
  mae: string;
  pai: string;
  telefone_principal: string;
  telefone_alternativo: string;
  outro_contacto: number;
  user_pessoa_id: number;
  
  //Funcionario
    base_pessoa_id: number;
    rh_categoria_id: number;
    base_area_id: number;
    estado: number;
  }